#216: rename Tx components to TokenTransfer
