+++ packages/template/src/views/Portfolio.vue
@@ -3,7 +3,7 @@
     <div class="row row-sm-revers">
       <div class="col-md-6">
         <SpAssets />
-        <SpTxList />
+        <SpTokenTransferList />
       </div>
       <div class="col-md-5 col-lg-4 col-md-offset-1 col-lg-offset-2">
         <SpTokenTransfer />
@@ -13,14 +13,14 @@
 </template>
 
 <script>
-import { SpAssets, SpTokenTransfer, SpTxList } from '@starport/vue'
+import { SpAssets, SpTokenTransfer, SpTokenTransferList } from '@starport/vue'
 import { computed } from 'vue'
 import { useStore } from 'vuex'
 
 export default {
   name: 'Portfolio',
 
-  components: { SpTokenTransfer, SpAssets, SpTxList },
+  components: { SpTokenTransfer, SpAssets, SpTokenTransferList },
 
   setup() {
     // store
