+++ packages/vue/src/components/SpGetTokenTransferList/index.ts
@@ -2,7 +2,7 @@ import { App as Application } from 'vue'
 
 import { registerComponent } from '../../utils/plugins/index'
 // @ts-ignore
-import C from './SpTxListItem.vue'
+import C from './SpGetTokenTransferList.vue'
 
 export const Plugin = {
   install(vue: Application): void {
