+++ packages/vue/src/components/SpTokenTransferListItem/index.ts
@@ -1,8 +1,8 @@
 import { App as Application } from 'vue'
 
-import { registerComponent } from './../../utils/plugins/index'
+import { registerComponent } from '../../utils/plugins/index'
 // @ts-ignore
-import C from './SpGetTxList.vue'
+import C from './SpTokenTransferListItem.vue'
 
 export const Plugin = {
   install(vue: Application): void {
