+++ packages/vue/src/components/index.ts
@@ -10,7 +10,7 @@ export { default as SpCrudDelete } from './SpCrudDelete'
 export { default as SpCrudRead } from './SpCrudRead'
 export { default as SpCrudUpdate } from './SpCrudUpdate'
 export { default as SpDropdown } from './SpDropdown'
-export { default as SpGetTxList } from './SpGetTxList'
+export { default as SpGetTokenTransferList } from './SpGetTokenTransferList'
 export { default as SpLinkIcon } from './SpLinkIcon'
 export { default as SpModal } from './SpModal'
 export { default as SpNavbar } from './SpNavbar'
@@ -18,5 +18,5 @@ export { default as SpSpacer } from './SpSpacer'
 export { default as SpSystemBar } from './SpSystemBar'
 export { default as SpTheme } from './SpTheme'
 export { default as SpTokenTransfer } from './SpTokenTransfer'
-export { default as SpTxList } from './SpTxList'
+export { default as SpTokenTransferList } from './SpTokenTransferList'
 export { default as SpTypography } from './SpTypography'
